requirements.txt
mysql-connector-python
pillow
tkcalendar
openpyxl
reportlab
tabulate 